Move get_all_ips to flocker.common
