+++ flocker/common/_net.py
@@ -1,0 +1,25 @@
+# Copyright Hybrid Logic Ltd.  See LICENSE file for details.
+
+"""
+Network utilities.
+"""
+
+import netifaces
+
+
+def get_all_ips():
+    """
+    Find all IPs for this machine.
+
+    :return: ``list`` of IP addresses (``bytes``).
+    """
+    ips = []
+    interfaces = netifaces.interfaces()
+    for interface in interfaces:
+        addresses = netifaces.ifaddresses(interface)
+        ipv4 = addresses.get(netifaces.AF_INET)
+        if not ipv4:
+            continue
+        for address in ipv4:
+            ips.append(address['addr'])
+    return ips
