+++ flocker/node/test/test_script.py
@@ -3,7 +3,6 @@
 """
 Tests for :module:`flocker.node.script`.
 """
-import netifaces
 import yaml
 
 from zope.interface.verify import verifyObject
@@ -15,6 +14,7 @@ from twisted.application.service import Service
 
 from ...volume.testtools import make_volume_options_tests
 from ...common.script import ICommandLineScript
+from ...common import get_all_ips
 
 from ..script import (
     ZFSAgentOptions, ZFSAgentScript, AgentScript, ContainerAgentOptions,
@@ -84,26 +84,6 @@ class ZFSAgentScriptTests(SynchronousTestCase):
                                            host=u"10.0.0.1",
                                            port=1234),
                           P2PManifestationDeployer, service, True))
-
-
-def get_all_ips():
-    """
-    Find all IPs for this machine.
-
-    XXX: Move this to flocker.common.???
-
-    :return: ``list`` of IP addresses (``bytes``).
-    """
-    ips = []
-    interfaces = netifaces.interfaces()
-    for interface in interfaces:
-        addresses = netifaces.ifaddresses(interface)
-        ipv4 = addresses.get(netifaces.AF_INET)
-        if not ipv4:
-            continue
-        for address in ipv4:
-            ips.append(address['addr'])
-    return ips
 
 
 class AgentServiceFactoryTests(SynchronousTestCase):
